+++ packages/core/src/payment/strategies/barclays/index.ts
@@ -1,1 +1,0 @@
-export { default as BarclaysPaymentStrategy } from './barclays-payment-strategy';