+++ packages/stripe-integration/src/index.ts
@@ -4,8 +4,6 @@ export { default as createStripeUPECustomerStrategy } from './stripe-upe/create-
 export { default as createStripeOCSPaymentStrategy } from './stripe-ocs/create-stripe-ocs-payment-strategy';
 export { default as createStripeLinkV2CustomerStrategy } from './stripe-ocs/create-stripe-link-v2-customer-strategy';
 
-export { default as StripeScriptLoader } from './stripev3/stripev3-script-loader';
-
 export { default as StripeV3PaymentStrategy } from './stripev3/stripev3-payment-strategy';
 export { default as StripeUPEPaymentStrategy } from './stripe-upe/stripe-upe-payment-strategy';
 export { default as StripeUPECustomerStrategy } from './stripe-upe/stripe-upe-customer-strategy';
@@ -28,5 +26,3 @@ export {
     default as StripeOCSPaymentInitializeOption,
     WithStripeOCSPaymentInitializeOptions,
 } from './stripe-ocs/stripe-ocs-initialize-options';
-
-export { StripeAppearanceOptions, StripeCustomFont } from './stripe-utils';