+++ packages/stripe-utils/src/stripe-initialize-options.ts
@@ -1,0 +1,17 @@
+import { StripeAppearanceValues } from './stripe';
+
+export default interface StripePaymentInitializeOptions {
+    /**
+     * The location to insert the credit card number form field.
+     */
+    containerId: string;
+
+    /**
+     * Checkout styles from store theme
+     */
+    style?: Record<string, StripeAppearanceValues>;
+
+    onError?(error?: Error): void;
+
+    render(): void;
+}